requirements.txt
# Testing and Development
pytest>=8.3.5
mock>=5.1.0
faker>=19.0.0

# Database Connectivity
cx_Oracle>=8.3.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.39

# Data Processing and Analysis
numpy>=1.26.4
pandas>=2.1.4
polars>=1.25.2
python-dateutil>=2.9.0
pytz>=2025.1
scikit-learn>=1.6.1

# Data Validation and Quality
great-expectations>=1.3.10
pandera>=0.23.1
ydata-profiling>=4.14.0

# Visualization and Documentation
plotly>=6.0.0
sadisplay>=0.4.9
graphviz>=0.20.1
Jinja2>=3.1.3

# String Processing
rapidfuzz>=3.12.2

# Package Management
pip>=25.0.1

# Jupyter Notebook
jupyter>=1.0.0
notebook>=7.0.0
ipynb-py-convert>=0.5.1

# Additional dependencies
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
